feature: simplifying home page code
